emplace Fix
Referring to P0088R0 (http://www.open-std.org/JTC1/SC22/WG21/docs/papers/2015/p0088r0.pdf), pages 20 & 21 one finds that for the overloads of emplace which take a size_t as their first template parameter the following postcondition is present: "index() is I".

The implementation was not respecting that, instead setting index() to the lowest N for which the type of the Nth parameter in the parameter pack upon which the variant was templated was equal to the type of the Ith such parameter.

To trivially reproduce this issue:

juice::variant<int,int> v;
v.emplace<1>(0);
std::cout << v.index() << std::endl;

Should print 1, instead prints 0.